Add `ignore_chars` parameter to `counts_tokens`.
This allows the user to ignore tokens in the count if they have a
character (like '°') that they use to annotate some of the tokens.
For example, the character might be used to indicate tokens that
the user does not require themselves to produce in order to get
the card 'right'.