Add `de_override_df` parameter to `create_de_additional_output`.
Add new `process_de_override_df` that processes the data frame, if
it is passed. Two additional parameters were also added to
`create_de_additional_output` to be passed through to this function.

Add code to create the data frame that will be passed in this new
parameter. The code is controlled by a parameter setting that is
currently off, so that the output still matches the previous output.