Support new `de_xref` column in `additional_input_notes.txt`.
The new columns be provide the headword for the merge with the de1 flagged
records, if it's populated. Otherwise, `de1` will be used.

Also, change `note_id` in `de_additional.txt` output file to ensure
uniqueness, remove mention of a specific language, and allow multiple
records in the `additional` file with the same headword. For records derived
from the flagged tokens in the main file, the headword is a component in
the id, so a 'HW_' is prepended to the headword (previously was 'DE1_').
For the records from the `additional_input_notes.txt` file, a 'AD_' is
prepended to the `id` column from this file. This was previously also
'DE1_' + the headword from the record.