Move tokenization out of `tag_audio_and_markup`.
In this function, rename `fields` parameter to `tokens` and remove `sep`
parameter.

In `example1.py`, create new `field_info` dictionary, where for each item the
key is the field name and the value is a namedtuple as the value that holds
the separator and whether the chapter should be assigned.

Calls to `de_additional.py` previously tokenized and retokenized using a
semi-colon separator since a separator was required to be passed and we knew
the test input data did not have semi-colons in `de1`. This has been removed
and `de1` is not tokenized at all.