Remove tokenization done inside `make_prompt_and_answer_table`.
The strings must now be tokenized before being passed to this function.
The parameters `tokenized_prompts` and `tokenized_answers` are now
renamed to `expr_prompts` and `expr_answers`, respectively, and
the `tokenized_omitted` in the returned value is now `exprs_omitted`.

The `de3` and `de3_prompts` parameters to `create_de_additional_output` are
now renamed to `de3_list` and `de3_prompts_list`, and the values passed
should be the result of tokenizing `de3` and `de3_prompts`.