Remove tokenization done in `make_new_cards`.
This was the primary aim of the other recent commits to pull out
the tokenization, as the tokenization being done inside the function
may not have been of wide interest to users, as the separator
character was dependent on `part_of_speech` and whether a certain
text string was in `de1`. `de1_list` and `de2_list` are added as
parameters to this function, and `de1` and `sep` are removed.