Rename `nl_hint` to `nl_abbr` and `tl_hint` to `tl_abbr` in `create_tl_additional_output`.
The reason is the old names are confusingly similar to `tl1_hint`, a free-text
hint the user creates, but the fields we renamed are meant to just be a very
short abbreviation to indicate in what language the answer will be in.

Also, rename `def_type` output variable to `answer_lang` so that it is more
descriptive.